+++ app/components/module/workPlace/ItemEdit.tsx
@@ -1,5 +1,5 @@
 // import type { ModuleItemResponse } from "api/models/ModuleItemResponse";
-// import { useUpdateItemMutation } from "api/hooks/moduleAndStageQueryHooks";
+// import { useUpdateItemMutation } from "api/hooks/moduleQueryHooks";
 // import { useEffect, useState } from "react";
 
 // interface ItemEditProps {
